+++ kalite/distributed/backup.py
@@ -27,27 +27,8 @@ def setup_backup():
     except KeyError:
         pass
 
-def restore_list():
-    file_list = os.listdir(backup_dirpath) #Retrieve the filenames present in the backup_dirpath
-    filelist = []
-    filenumber = 0
-
-    for eachfile in file_list:
-        filelist.append(eachfile) #Store files one by one
-        print filelist.index(eachfile), eachfile #show file number: index
-        filenumber += 1
-
-    print "Please enter file-number to restore:\n"
-
-    backup_option = int(raw_input())
-    restoref = filelist[backup_option] #file to be restored based on x
-    backup_filepath = os.path.join(backup_dirpath, restoref) #create file path
-    call_command('dbrestore', filepath=backup_filepath, database='default') #perform restore
-
 def rotate_backups():
     now = time.time()
     for fp in os.listdir(backup_dirpath):
         if os.stat(os.path.join(backup_dirpath, fp)).st_mtime < now - 7 * 86400:
             os.remove(os.path.join(backup_dirpath, fp))
-
-setup_backup()
